+++ test/test_catalog.py
@@ -1,13 +1,15 @@
 from __future__ import print_function
 
 import unittest
+
 from test.pattern_for_tests import *
+import pystitch
 
 
 class TestDataCatalog(unittest.TestCase):
 
     def test_catalog_files(self):
-        for f in EmbPattern.supported_formats():
+        for f in pystitch.supported_formats():
             self.assertIn("extensions", f)
             self.assertIn("extension", f)
             self.assertIn("description", f)
